+++ projects/ontimize-web-ngx/src/lib/components/app-sidenav/index.ts
@@ -1,6 +1,7 @@
 export * from './image/o-app-sidenav-image.component';
 export * from './menu-group/o-app-sidenav-menu-group.component';
 export * from './menu-item/o-app-sidenav-menu-item.component';
+export * from './menu-section/o-app-sidenav-menu-section.component';
 export * from './o-app-sidenav.component';
 export * from './o-app-sidenav.module';
 export * from './o-app-sidenav-base.class';