+++ projects/ontimize-web-ngx/src/lib/components/bar-menu/menu-nested/o-bar-menu-nested.component.ts
@@ -1,6 +1,7 @@
 import { ChangeDetectionStrategy, Component, Injector } from '@angular/core';
 import { NgClass } from '@angular/common';
 
+import { MenuSection } from '../../../interfaces/app-menu.interface';
 import { AppMenuService } from '../../../services/app-menu.service';
 import { MenuRootItem } from '../../../types/menu-root-item.type';
 import { OBarMenuGroupComponent } from '../menu-group/o-bar-menu-group.component';
@@ -28,6 +29,20 @@ export class OBarMenuNestedComponent {
     protected injector: Injector
   ) {
     this.appMenuService = this.injector.get(AppMenuService);
+  }
+
+  /**
+   * Menu sections have no equivalent in the bar menu, so they are rendered as regular menu groups.
+   */
+  getItemType(item: MenuRootItem): string {
+    const type = this.appMenuService.getMenuItemType(item);
+    return type === 'section' ? 'group' : type;
+  }
+
+  isItemVisible(item: MenuRootItem): boolean {
+    return this.appMenuService.isMenuSection(item)
+      ? this.appMenuService.isSectionVisible(item as MenuSection)
+      : this.appMenuService.isVisible(item);
   }
 
 }